+++ src/OsType/ValueBuilders/BuildTypeFromEtcIssue.php
@@ -163,6 +163,7 @@ class BuildTypeFromEtcIssue implements BuildTypeFromFile
     private static $osTypes = [
         "|^CentOS release (?<version>\d+\.\d+)|" => CentOS::class,
         "|^Debian GNU/Linux (?<version>\d+)|" => Debian::class,
+        "|^Linux Mint (?<version>\d+\.\d+)|" => LinuxMint::class,
         "|^Ubuntu (?<version>\d+\.\d+)|" => Ubuntu::class,
     ];
 }