+++ tests/OsType/ValueBuilders/BuildTypeFromLsbReleaseTest.php
@@ -295,6 +295,10 @@ class BuildTypeFromLsbReleaseTest extends PHPUnit_Framework_TestCase
                 new Debian('8.2')
             ],
             [
+                __DIR__ . '/lsb_release-examples/linuxmint-17.3.php',
+                new LinuxMint('17.3')
+            ],
+            [
                 __DIR__ . '/lsb_release-examples/ubuntu-15.04.php',
                 new Ubuntu('15.04')
             ]