+++ tests/OsType/ValueBuilders/lsb_release-examples/debian-8.2.php
@@ -1,0 +1,10 @@
+#!/usr/bin/env php
+<?php
+
+echo <<<EOS
+Distributor ID: Debian
+Description:    Debian GNU/Linux 8.2 (jessie)
+Release:    8.2
+Codename:   jessie
+
+EOS;