+++ tests/OsType/ValueBuilders/lsb_release-examples/linuxmint-17.3.php
@@ -1,0 +1,10 @@
+#!/usr/bin/env php
+<?php
+
+echo <<<EOS
+Distributor ID: LinuxMint
+Description:    Linux Mint 17.3 Rosa
+Release:    17.3
+Codename:   rosa
+
+EOS;