+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 Nothing yet.
 
+## 1.1.0 - Mon Jan 17 2016
+
+### New
+
+* Support for Debian 8.x (aka Jessie) added.
+* Support for LinuxMint 17.3 (aka Rosa) added.
+
 ## 1.0.0 - Sun Jan 16 2016
 
 ### New